requirements.txt
asgiref==3.4.1
cffi==1.14.6
Django==3.2.5
djangorestframework==3.12.4
Pillow==8.3.1
pkg-resources==0.0.0
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
